+++ pages/maps.php
@@ -76,49 +76,49 @@ For support and installation notes visit http://www.hlxcommunity.com
 			(
 				'kills',
 				'Kills',
-				'width=8&align=right'
+				'width=8&align=center'
 			),
 			new TableColumn
 			(
 				'kpercent',
 				'%',
-				'width=7&sort=no&align=right&append=' . urlencode('%')
+				'width=7&sort=no&align=center&append=' . urlencode('%')
 			),
 			new TableColumn
 			(
 				'kpercent',
 				'Ratio',
-				'width=16&sort=no&type=bargraph'
+				'width=16&sort=no&type=bargraph&align=middle'
 			),
 			new TableColumn
 			(
 				'headshots',
 				'Headshots',
-				'width=8&align=right'
+				'width=8&align=center'
 			),
 			new TableColumn
 			(
 				'hpercent',
 				'%',
-				'width=7&sort=no&align=right&append=' . urlencode('%')
+				'width=7&sort=no&align=center&append=' . urlencode('%')
 			),
 			new TableColumn
 			(
 				'hpercent',
 				'Ratio',
-				'width=16&sort=no&type=bargraph'
+				'width=16&sort=no&type=bargraph&align=middle'
 			),
 			new TableColumn
 			(
 				'hpk',
 				'HS:K',
-				'width=9&align=right'
+				'width=9&align=center'
 			),
 			new TableColumn
 			(
 				'map',
 				'HeatMap',
-				'width=4&type=heatmap'
+				'width=4&type=heatmap&align=center'
 			)
 		),
 		'map',