+++ pages/playerinfo_servers.php
@@ -55,50 +55,50 @@ For support and installation notes visit http://www.hlxcommunity.com
 			(
 				'kills',
 				'Kills',
-				'width=6&align=right'
+				'width=6&align=center'
 			),
 			new TableColumn
 			(
 				'kpercent',
 				'%',
-				'width=5&sort=no&align=right&append=' . urlencode('%')
+				'width=5&sort=no&align=center&append=' . urlencode('%')
 			),
 			new TableColumn
 			(
 				'kpercent',
 				'Ratio',
-				'width=15&sort=no&type=bargraph'
+				'width=15&sort=no&type=bargraph&align=middle'
 			),
 			new TableColumn
 			(
 				'deaths',
 				'Deaths',
-				'width=6&align=right'
+				'width=6&align=center'
 			),
 			new TableColumn(
 				'kpd',
 				'K:D',
-				'width=5&align=right'
+				'width=5&align=center'
 			),
 			new TableColumn(
 				'headshots',
 				'Headshots',
-				'width=6&align=right'
+				'width=6&align=center'
 			),
 			new TableColumn(
 				'hpercent',
 				'Percentage of Headshots',
-				'width=16&sort=no&type=bargraph'
+				'width=16&sort=no&type=bargraph&align=middle'
 			),
 			new TableColumn(
 				'hpercent',
 				'%',
-				'width=5&sort=no&align=right&append=' . urlencode('%')
+				'width=5&sort=no&align=center&append=' . urlencode('%')
 			),
 			new TableColumn(
 				'hpk',
 				'HS:K',
-				'width=5&align=right'
+				'width=5&align=center'
 			)
 			
 		),